+++ src/internal/connector/data_collections_test.go
@@ -10,11 +10,8 @@ import (
 
 	"github.com/alcionai/corso/src/internal/connector/exchange"
 	"github.com/alcionai/corso/src/internal/connector/sharepoint"
-	"github.com/alcionai/corso/src/internal/connector/support"
-	"github.com/alcionai/corso/src/internal/data"
 	"github.com/alcionai/corso/src/internal/tester"
 	"github.com/alcionai/corso/src/pkg/control"
-	"github.com/alcionai/corso/src/pkg/path"
 	"github.com/alcionai/corso/src/pkg/selectors"
 )
 
@@ -105,7 +102,14 @@ func (suite *ConnectorDataCollectionIntegrationSuite) TestExchangeDataCollection
 
 	for _, test := range tests {
 		suite.T().Run(test.name, func(t *testing.T) {
-			collection, err := connector.ExchangeDataCollection(ctx, test.getSelector(t), nil, control.Options{})
+			collection, err := exchange.DataCollection(
+				ctx,
+				test.getSelector(t),
+				nil,
+				[]string{suite.user},
+				connector.credentials,
+				connector.UpdateStatus,
+				control.Options{})
 			require.NoError(t, err)
 			// Categories with delta endpoints will produce a collection for metadata
 			// as well as the actual data pulled.
@@ -229,354 +233,6 @@ func (suite *ConnectorDataCollectionIntegrationSuite) TestSharePointDataCollecti
 			assert.NotZero(t, status.Successful)
 
 			t.Log(status.String())
-		})
-	}
-}
-
-// ---------------------------------------------------------------------------
-// CreateExchangeCollection tests
-// ---------------------------------------------------------------------------
-
-type ConnectorCreateExchangeCollectionIntegrationSuite struct {
-	suite.Suite
-	connector *GraphConnector
-	user      string
-	site      string
-}
-
-func TestConnectorCreateExchangeCollectionIntegrationSuite(t *testing.T) {
-	if err := tester.RunOnAny(
-		tester.CorsoCITests,
-		tester.CorsoConnectorCreateExchangeCollectionTests,
-	); err != nil {
-		t.Skip(err)
-	}
-
-	suite.Run(t, new(ConnectorCreateExchangeCollectionIntegrationSuite))
-}
-
-func (suite *ConnectorCreateExchangeCollectionIntegrationSuite) SetupSuite() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	_, err := tester.GetRequiredEnvVars(tester.M365AcctCredEnvs...)
-	require.NoError(suite.T(), err)
-	suite.connector = loadConnector(ctx, suite.T(), Users)
-	suite.user = tester.M365UserID(suite.T())
-	suite.site = tester.M365SiteID(suite.T())
-	tester.LogTimeOfTest(suite.T())
-}
-
-func (suite *ConnectorCreateExchangeCollectionIntegrationSuite) TestMailFetch() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	var (
-		t      = suite.T()
-		userID = tester.M365UserID(t)
-	)
-
-	tests := []struct {
-		name        string
-		scope       selectors.ExchangeScope
-		folderNames map[string]struct{}
-	}{
-		{
-			name: "Folder Iterative Check Mail",
-			scope: selectors.NewExchangeBackup().MailFolders(
-				[]string{userID},
-				[]string{exchange.DefaultMailFolder},
-				selectors.PrefixMatch(),
-			)[0],
-			folderNames: map[string]struct{}{
-				exchange.DefaultMailFolder: {},
-			},
-		},
-	}
-
-	gc := loadConnector(ctx, t, Users)
-
-	for _, test := range tests {
-		suite.T().Run(test.name, func(t *testing.T) {
-			collections, err := gc.createExchangeCollections(ctx, test.scope, exchange.DeltaPaths{}, control.Options{})
-			require.NoError(t, err)
-
-			for _, c := range collections {
-				if c.FullPath().Service() == path.ExchangeMetadataService {
-					continue
-				}
-
-				require.NotEmpty(t, c.FullPath().Folder())
-				folder := c.FullPath().Folder()
-
-				delete(test.folderNames, folder)
-			}
-
-			assert.Empty(t, test.folderNames)
-		})
-	}
-}
-
-func (suite *ConnectorCreateExchangeCollectionIntegrationSuite) TestDelta() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	var (
-		userID = tester.M365UserID(suite.T())
-		gc     = loadConnector(ctx, suite.T(), Users)
-	)
-
-	tests := []struct {
-		name  string
-		scope selectors.ExchangeScope
-	}{
-		{
-			name: "Mail",
-			scope: selectors.NewExchangeBackup().MailFolders(
-				[]string{userID},
-				[]string{exchange.DefaultMailFolder},
-				selectors.PrefixMatch(),
-			)[0],
-		},
-		{
-			name: "Contacts",
-			scope: selectors.NewExchangeBackup().ContactFolders(
-				[]string{userID},
-				[]string{exchange.DefaultContactFolder},
-				selectors.PrefixMatch(),
-			)[0],
-		},
-	}
-	for _, test := range tests {
-		suite.T().Run(test.name, func(t *testing.T) {
-			// get collections without providing any delta history (ie: full backup)
-			collections, err := gc.createExchangeCollections(ctx, test.scope, exchange.DeltaPaths{}, control.Options{})
-			require.NoError(t, err)
-			assert.Less(t, 1, len(collections), "retrieved metadata and data collections")
-
-			var metadata data.Collection
-
-			for _, coll := range collections {
-				if coll.FullPath().Service() == path.ExchangeMetadataService {
-					metadata = coll
-				}
-			}
-
-			require.NotNil(t, metadata, "collections contains a metadata collection")
-
-			cdps, err := exchange.ParseMetadataCollections(ctx, []data.Collection{metadata})
-			require.NoError(t, err)
-
-			dps := cdps[test.scope.Category().PathType()]
-
-			// now do another backup with the previous delta tokens,
-			// which should only contain the difference.
-			collections, err = gc.createExchangeCollections(ctx, test.scope, dps, control.Options{})
-			require.NoError(t, err)
-
-			// TODO(keepers): this isn't a very useful test at the moment.  It needs to
-			// investigate the items in the original and delta collections to at least
-			// assert some minimum assumptions, such as "deltas should retrieve fewer items".
-			// Delta usage is commented out at the moment, anyway.  So this is currently
-			// a sanity check that the minimum behavior won't break.
-			for _, coll := range collections {
-				if coll.FullPath().Service() != path.ExchangeMetadataService {
-					ec, ok := coll.(*exchange.Collection)
-					require.True(t, ok, "collection is *exchange.Collection")
-					assert.NotNil(t, ec)
-				}
-			}
-		})
-	}
-}
-
-// TestMailSerializationRegression verifies that all mail data stored in the
-// test account can be successfully downloaded into bytes and restored into
-// M365 mail objects
-func (suite *ConnectorCreateExchangeCollectionIntegrationSuite) TestMailSerializationRegression() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	var (
-		t         = suite.T()
-		connector = loadConnector(ctx, t, Users)
-		sel       = selectors.NewExchangeBackup()
-	)
-
-	sel.Include(sel.MailFolders([]string{suite.user}, []string{exchange.DefaultMailFolder}, selectors.PrefixMatch()))
-
-	collection, err := connector.createExchangeCollections(ctx, sel.Scopes()[0], exchange.DeltaPaths{}, control.Options{})
-	require.NoError(t, err)
-
-	for _, edc := range collection {
-		suite.T().Run(edc.FullPath().String(), func(t *testing.T) {
-			streamChannel := edc.Items()
-			// Verify that each message can be restored
-			for stream := range streamChannel {
-				buf := &bytes.Buffer{}
-
-				read, err := buf.ReadFrom(stream.ToReader())
-				assert.NoError(t, err)
-				assert.NotZero(t, read)
-
-				message, err := support.CreateMessageFromBytes(buf.Bytes())
-				assert.NotNil(t, message)
-				assert.NoError(t, err)
-			}
-		})
-	}
-
-	status := connector.AwaitStatus()
-	suite.NotNil(status)
-	suite.Equal(status.ObjectCount, status.Successful)
-}
-
-// TestContactSerializationRegression verifies ability to query contact items
-// and to store contact within Collection. Downloaded contacts are run through
-// a regression test to ensure that downloaded items can be uploaded.
-func (suite *ConnectorCreateExchangeCollectionIntegrationSuite) TestContactSerializationRegression() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	connector := loadConnector(ctx, suite.T(), Users)
-
-	tests := []struct {
-		name          string
-		getCollection func(t *testing.T) []data.Collection
-	}{
-		{
-			name: "Default Contact Folder",
-			getCollection: func(t *testing.T) []data.Collection {
-				scope := selectors.
-					NewExchangeBackup().
-					ContactFolders([]string{suite.user}, []string{exchange.DefaultContactFolder}, selectors.PrefixMatch())[0]
-				collections, err := connector.createExchangeCollections(ctx, scope, exchange.DeltaPaths{}, control.Options{})
-				require.NoError(t, err)
-
-				return collections
-			},
-		},
-	}
-
-	for _, test := range tests {
-		suite.T().Run(test.name, func(t *testing.T) {
-			edcs := test.getCollection(t)
-			require.GreaterOrEqual(t, len(edcs), 1, "expected 1 <= num collections <= 2")
-			require.GreaterOrEqual(t, 2, len(edcs), "expected 1 <= num collections <= 2")
-
-			for _, edc := range edcs {
-				isMetadata := edc.FullPath().Service() == path.ExchangeMetadataService
-				count := 0
-
-				for stream := range edc.Items() {
-					buf := &bytes.Buffer{}
-					read, err := buf.ReadFrom(stream.ToReader())
-					assert.NoError(t, err)
-					assert.NotZero(t, read)
-
-					if isMetadata {
-						continue
-					}
-
-					contact, err := support.CreateContactFromBytes(buf.Bytes())
-					assert.NotNil(t, contact)
-					assert.NoError(t, err, "error on converting contact bytes: "+buf.String())
-					count++
-				}
-
-				if isMetadata {
-					continue
-				}
-
-				assert.Equal(t, edc.FullPath().Folder(), exchange.DefaultContactFolder)
-				assert.NotZero(t, count)
-			}
-
-			status := connector.AwaitStatus()
-			suite.NotNil(status)
-			suite.Equal(status.ObjectCount, status.Successful)
-		})
-	}
-}
-
-// TestEventsSerializationRegression ensures functionality of createCollections
-// to be able to successfully query, download and restore event objects
-func (suite *ConnectorCreateExchangeCollectionIntegrationSuite) TestEventsSerializationRegression() {
-	ctx, flush := tester.NewContext()
-	defer flush()
-
-	connector := loadConnector(ctx, suite.T(), Users)
-
-	tests := []struct {
-		name, expected string
-		getCollection  func(t *testing.T) []data.Collection
-	}{
-		{
-			name:     "Default Event Calendar",
-			expected: exchange.DefaultCalendar,
-			getCollection: func(t *testing.T) []data.Collection {
-				sel := selectors.NewExchangeBackup()
-				sel.Include(sel.EventCalendars([]string{suite.user}, []string{exchange.DefaultCalendar}, selectors.PrefixMatch()))
-
-				collections, err := connector.createExchangeCollections(
-					ctx,
-					sel.Scopes()[0],
-					exchange.DeltaPaths{},
-					control.Options{})
-				require.NoError(t, err)
-
-				return collections
-			},
-		},
-		{
-			name:     "Birthday Calendar",
-			expected: "Birthdays",
-			getCollection: func(t *testing.T) []data.Collection {
-				sel := selectors.NewExchangeBackup()
-				sel.Include(sel.EventCalendars([]string{suite.user}, []string{"Birthdays"}, selectors.PrefixMatch()))
-
-				collections, err := connector.createExchangeCollections(
-					ctx,
-					sel.Scopes()[0],
-					exchange.DeltaPaths{},
-					control.Options{})
-				require.NoError(t, err)
-
-				return collections
-			},
-		},
-	}
-
-	for _, test := range tests {
-		suite.T().Run(test.name, func(t *testing.T) {
-			collections := test.getCollection(t)
-			require.Equal(t, len(collections), 2)
-
-			for _, edc := range collections {
-				if edc.FullPath().Service() != path.ExchangeMetadataService {
-					assert.Equal(t, test.expected, edc.FullPath().Folder())
-				} else {
-					assert.Equal(t, "", edc.FullPath().Folder())
-				}
-
-				streamChannel := edc.Items()
-
-				for stream := range streamChannel {
-					buf := &bytes.Buffer{}
-
-					read, err := buf.ReadFrom(stream.ToReader())
-					assert.NoError(t, err)
-					assert.NotZero(t, read)
-
-					event, err := support.CreateEventFromBytes(buf.Bytes())
-					assert.NotNil(t, event)
-					assert.NoError(t, err, "creating event from bytes: "+buf.String())
-				}
-			}
-
-			status := connector.AwaitStatus()
-			suite.NotNil(status)
-			suite.Equal(status.ObjectCount, status.Successful)
 		})
 	}
 }
