+++ src/internal/connector/exchange/data_collections.go
@@ -3,12 +3,19 @@ package exchange
 import (
 	"context"
 	"encoding/json"
+	"fmt"
 
+	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 
 	"github.com/alcionai/corso/src/internal/connector/graph"
+	"github.com/alcionai/corso/src/internal/connector/support"
 	"github.com/alcionai/corso/src/internal/data"
+	"github.com/alcionai/corso/src/internal/observe"
+	"github.com/alcionai/corso/src/pkg/account"
+	"github.com/alcionai/corso/src/pkg/control"
 	"github.com/alcionai/corso/src/pkg/path"
+	"github.com/alcionai/corso/src/pkg/selectors"
 )
 
 // MetadataFileNames produces the category-specific set of filenames used to
@@ -53,7 +60,7 @@ type DeltaPath struct {
 
 // ParseMetadataCollections produces a map of structs holding delta
 // and path lookup maps.
-func ParseMetadataCollections(
+func parseMetadataCollections(
 	ctx context.Context,
 	colls []data.Collection,
 ) (CatDeltaPaths, error) {
@@ -145,4 +152,112 @@ func ParseMetadataCollections(
 	}
 
 	return cdp, nil
+}
+
+// DataCollections returns a DataCollection which the caller can
+// use to read mailbox data out for the specified user
+// Assumption: User exists
+//
+//	Add iota to this call -> mail, contacts, calendar,  etc.
+func DataCollection(
+	ctx context.Context,
+	selector selectors.Selector,
+	metadata []data.Collection,
+	userPNs []string,
+	acct account.M365Config,
+	su support.StatusUpdater,
+	ctrlOpts control.Options,
+) ([]data.Collection, error) {
+	eb, err := selector.ToExchangeBackup()
+	if err != nil {
+		return nil, errors.Wrap(err, "exchangeDataCollection: parsing selector")
+	}
+
+	var (
+		scopes      = eb.DiscreteScopes(userPNs)
+		collections = []data.Collection{}
+		errs        error
+	)
+
+	cdps, err := parseMetadataCollections(ctx, metadata)
+	if err != nil {
+		return nil, err
+	}
+
+	for _, scope := range scopes {
+		dps := cdps[scope.Category().PathType()]
+
+		dcs, err := createCollections(
+			ctx,
+			acct,
+			scope,
+			dps,
+			control.Options{},
+			su)
+		if err != nil {
+			user := scope.Get(selectors.ExchangeUser)
+			return nil, support.WrapAndAppend(user[0], err, errs)
+		}
+
+		collections = append(collections, dcs...)
+	}
+
+	return collections, errs
+}
+
+// createCollections - utility function that retrieves M365
+// IDs through Microsoft Graph API. The selectors.ExchangeScope
+// determines the type of collections that are retrieved.
+func createCollections(
+	ctx context.Context,
+	acct account.M365Config,
+	scope selectors.ExchangeScope,
+	dps deltaPaths,
+	ctrlOpts control.Options,
+	su support.StatusUpdater,
+) ([]data.Collection, error) {
+	var (
+		errs           *multierror.Error
+		users          = scope.Get(selectors.ExchangeUser)
+		allCollections = make([]data.Collection, 0)
+	)
+
+	// Create collection of ExchangeDataCollection
+	for _, user := range users {
+		collections := make(map[string]data.Collection)
+
+		qp := graph.QueryParams{
+			Category:      scope.Category().PathType(),
+			ResourceOwner: user,
+			Credentials:   acct,
+		}
+
+		foldersComplete, closer := observe.MessageWithCompletion(fmt.Sprintf("∙ %s - %s:", qp.Category, user))
+		defer closer()
+		defer close(foldersComplete)
+
+		resolver, err := populateExchangeContainerResolver(ctx, qp)
+		if err != nil {
+			return nil, errors.Wrap(err, "getting folder cache")
+		}
+
+		err = filterContainersAndFillCollections(
+			ctx,
+			qp,
+			collections,
+			su,
+			resolver,
+			scope,
+			dps,
+			ctrlOpts)
+
+		if err != nil {
+			return nil, errors.Wrap(err, "filling collections")
+		}
+
+		foldersComplete <- struct{}{}
+		//TODO fillAllCollectiton
+	}
+
+	return allCollections, errs.ErrorOrNil()
 }
