+++ src/internal/connector/exchange/service_iterators.go
@@ -39,19 +39,19 @@ func hasErrorCode(err error, code string) bool {
 		*oDataError.GetError().GetCode() == code
 }
 
-// FilterContainersAndFillCollections is a utility function
+// filterContainersAndFillCollections is a utility function
 // that places the M365 object ids belonging to specific directories
 // into a Collection. Messages outside of those directories are omitted.
 // @param collection is filled with during this function.
 // Supports all exchange applications: Contacts, Events, and Mail
-func FilterContainersAndFillCollections(
+func filterContainersAndFillCollections(
 	ctx context.Context,
 	qp graph.QueryParams,
 	collections map[string]data.Collection,
 	statusUpdater support.StatusUpdater,
 	resolver graph.ContainerResolver,
 	scope selectors.ExchangeScope,
-	dps DeltaPaths,
+	dps deltaPaths,
 	ctrlOpts control.Options,
 ) error {
 	var (
