+++ src/internal/connector/support/status.go
@@ -20,7 +20,7 @@ type ConnectorOperationStatus struct {
 	ObjectCount       int
 	FolderCount       int
 	Successful        int
-	errorCount        int
+	ErrorCount        int
 	incomplete        bool
 	incompleteReason  string
 	additionalDetails string
@@ -69,14 +69,14 @@ func CreateStatus(
 		ObjectCount:       cm.Objects,
 		FolderCount:       folders,
 		Successful:        cm.Successes,
-		errorCount:        numErr,
+		ErrorCount:        numErr,
 		incomplete:        hasErrors,
 		incompleteReason:  reason,
 		bytes:             cm.TotalBytes,
 		additionalDetails: details,
 	}
 
-	if status.ObjectCount != status.errorCount+status.Successful {
+	if status.ObjectCount != status.ErrorCount+status.Successful {
 		logger.Ctx(ctx).Errorw(
 			"status object count does not match errors + successes",
 			"objects", cm.Objects,
@@ -114,7 +114,7 @@ func MergeStatus(one, two ConnectorOperationStatus) ConnectorOperationStatus {
 		ObjectCount:       one.ObjectCount + two.ObjectCount,
 		FolderCount:       one.FolderCount + two.FolderCount,
 		Successful:        one.Successful + two.Successful,
-		errorCount:        one.errorCount + two.errorCount,
+		ErrorCount:        one.ErrorCount + two.ErrorCount,
 		bytes:             one.bytes + two.bytes,
 		incomplete:        hasErrors,
 		incompleteReason:  one.incompleteReason + ", " + two.incompleteReason,
